Move exported session types to type folder
